Add files via upload
This lesson is a thorough one which dives deep into the process of data cleaning and Preparation. The IMDB movies data set is utilized in this use case. It's a good lesson for learning and reinforcement of the tools, skills and process involved in Data Preparation and Cleaning. It will teach you how to write some simple and advanced functions that makes the process simpler, faster and efficient. These and more are what will be unveiled in this lesson. Remember to check the part 2 for the Data visualization, exploration and Analysis part.

And it supports jupyter notebook 100%.